introduce version of local repo detection using installed git
This is "slower" by a factor of 50x, but really only takes about 4.2ms
on my machine, and we're going to do a network call anyhow right after
that will take at least probably 100ms, so does it matter?

The primary motivation would be to reduce the binary size and have one
less dependency to deal with. This means git would have to be installed
on the local users machine, but for the usage of this particular CLI
tool, that's almost a certainty, and we're not likely to be operating
within an alpine environment. Switching over and eliminating go-git
would drop us from 16mb to 12mb. People probably dont care about binary
size though(?), so I think this is more academic in terms of what I
prefer, so it was an interesting comparison to do.

Comparison of each function added in comments, leaving in goGit as the
default for now, will decide and clean up once the program is feature
complete.