handle determining all relevant recent releases
Lots of attempting to deal with the discrepancies between how GitHub
releases work in practice, and what semantic versioning says about
pre-releases.

As part of this, we need to query enough data to know multiple releases,
since a) the GitHub API does not return releases flagged as prerelease
in its UI as part of getLatest, b) if we are on a prerelease version,
we need to also be able to know the last FULL release to generate a more
helpful changelog.